Force Metro to ignore some modules
curve25519-scalarmult-wasm and crypto are not used in runtime when this
module runs directly inside React Native, but Metro doesn't know how to
ignore them during bundling, *unless* we opt-out like this in package.json
in the "react-native" field. Note: this hack does not apply in nodejs-mobile
and it shouldn't.